# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aura Admin API",
    "description": "TCP + bearer-token surface: config, jobs, cron, memory, traces, skills, tools, channels, LLM.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "description": "All jobs currently tracked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Envelope for list endpoints. Lets us add `next_cursor`, `total`,\netc. later without breaking clients that parse `items`.",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "Mirror of [`aura_job::Job`].",
                        "required": [
                          "id",
                          "session_id",
                          "kind",
                          "status",
                          "created_at"
                        ],
                        "properties": {
                          "completed_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "error": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "id": {
                            "type": "string"
                          },
                          "input": {
                            "type": [
                              "object",
                              "null"
                            ]
                          },
                          "kind": {
                            "$ref": "#/components/schemas/OperationKind"
                          },
                          "output": {
                            "type": [
                              "object",
                              "null"
                            ]
                          },
                          "parent_job_id": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "session_id": {
                            "type": "string"
                          },
                          "started_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "status": {
                            "$ref": "#/components/schemas/JobStatus"
                          },
                          "trace_span_id": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Job store error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled job record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Cancel failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/memory": {
      "get": {
        "tags": [
          "memory"
        ],
        "operationId": "list_memory",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching memory entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Envelope for list endpoints. Lets us add `next_cursor`, `total`,\netc. later without breaking clients that parse `items`.",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "Mirror of [`aura_model::MemoryEntry`].",
                        "required": [
                          "id",
                          "user_id",
                          "content",
                          "category",
                          "importance",
                          "created_at",
                          "last_accessed"
                        ],
                        "properties": {
                          "category": {
                            "$ref": "#/components/schemas/MemoryCategory"
                          },
                          "content": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "embedding": {
                            "type": [
                              "array",
                              "null"
                            ],
                            "items": {
                              "type": "number",
                              "format": "float"
                            }
                          },
                          "expires_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "id": {
                            "type": "string"
                          },
                          "importance": {
                            "type": "number",
                            "format": "float"
                          },
                          "last_accessed": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "source_session_id": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "user_id": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Memory store error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "memory"
        ],
        "operationId": "store_memory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stored memory entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryEntry"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Memory store error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory/{id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "operationId": "delete_memory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory entry id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Memory entry deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Memory store error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/traces/{session_id}": {
      "get": {
        "tags": [
          "traces"
        ],
        "operationId": "get_trace",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session id whose trace to export",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw trace tree for the session. Shape mirrors `aura_trace::TraceNode` but is emitted as untyped JSON to keep the admin surface decoupled from internal trace crate changes.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No trace for that session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Uniform error envelope for every non-2xx admin response.\n\n`GatewayError::into_response` already emits `{\"error\": \"...\"}`; this\ntype just pins the shape in the spec so clients can discriminate on\nit rather than on status codes alone.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Job": {
        "type": "object",
        "description": "Mirror of [`aura_job::Job`].",
        "required": [
          "id",
          "session_id",
          "kind",
          "status",
          "created_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "input": {
            "type": [
              "object",
              "null"
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "output": {
            "type": [
              "object",
              "null"
            ]
          },
          "parent_job_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_id": {
            "type": "string"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "trace_span_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "description": "Mirror of [`aura_job::JobStatus`].",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "submitted",
          "accepted",
          "failed",
          "stuck"
        ]
      },
...
      "MemoryCategory": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "UserPreference",
              "KeyFact"
            ]
          }
        }
      },
      "MemoryEntry": {
        "type": "object",
        "description": "Mirror of [`aura_model::MemoryEntry`].",
        "required": [
          "id",
          "user_id",
          "content",
          "category",
          "importance",
          "created_at",
          "last_accessed"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/MemoryCategory"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "embedding": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "importance": {
            "type": "number",
            "format": "float"
          },
          "last_accessed": {
            "type": "string",
            "format": "date-time"
          },
          "source_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": "string"
          }
        }
      },
...
      "OperationKind": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "model",
              "type"
            ],
            "properties": {
              "model": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "llm_call"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "tool_name",
              "type"
            ],
            "properties": {
              "tool_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "tool_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "skill_name",
              "type"
            ],
            "properties": {
              "skill_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "skill_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "cron_job_id",
              "type"
            ],
            "properties": {
              "cron_job_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "cron_execution"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "strategy",
              "type"
            ],
            "properties": {
              "strategy": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "context_compression"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "operation",
              "type"
            ],
            "properties": {
              "operation": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "memory_operation"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "session_id",
              "type"
            ],
            "properties": {
              "session_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "user_message_handling"
                ]
              }
            }
          }
        ],
        "description": "Mirror of [`aura_job::OperationKind`]."
      },
...
      "StoreMemoryRequest": {
        "type": "object",
        "description": "`POST /v1/memory` body.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "importance": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}